trosat.cds: (re)move utility functions
